Initial implementation of plugin updates
